xfixes: move it under the Xext/ directory
