+++ Xext/xfixes/cursor.c
