+++ Xext/xfixes/disconnect.c
