+++ test/fixes.c
@@ -29,8 +29,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <X11/X.h>
-#include <xfixesint.h>
 #include <X11/extensions/xfixeswire.h>
+
+#include "Xext/xfixes/xfixesint.h"
 
 #include "tests-common.h"
